cluster: remove separate communication_v2 module
This commit moves the `create_sockets` implementation from
`mz_cluster::communication_v2` into `mz_cluster::communication` and
deletes the former. Pure code movement.
