+++ src/cluster/src/lib.rs
@@ -13,4 +13,3 @@
 
 pub mod client;
 pub mod communication;
-pub mod communication_v2;
